misc fixes
- fix forced-loop ambient_generics not working sometimes (gf_greybox)
- fix low env_beam lifetimes interpreted as infinite life and causing beam overflows
- fix robo grunts not exploding when gibbed
- fix an error message

+++ dlls/util.cpp
@@ -3877,7 +3877,7 @@ void MESSAGE_END() {
 bool UTIL_isSafeEntIndex(int idx, const char* action) {
 	if (sv_max_client_edicts && idx >= sv_max_client_edicts->value) {
 		ALERT(at_error, "Can't %s for edict %d '%s' (sv_max_client_edicts = %d)\n",
-			action, idx, STRING(INDEXENT(idx)->v.classname), sv_max_client_edicts->value);
+			action, idx, STRING(INDEXENT(idx)->v.classname), (int)sv_max_client_edicts->value);
 		return false;
 	}
